% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tsSetFunctions.R
\name{deactivateTsSet}
\alias{deactivateTsSet}
\title{Activate a Set of Time Series}
\usage{
deactivateTsSet(con, set_name, user_name = Sys.info()["user"],
  tbl = "timeseries_sets")
}
\arguments{
\item{con}{PostgreSQL connection object}

\item{user_name}{character name of the user. Defaults to system user.}

\item{tbl}{character name of set tqble. Defaults to timeseries\_sets.}
}
\description{
Activate a set of time series to get in the user's sight.
Deactivated sets are not deleted though.
}
\author{
Matthias Bannert, Ioan Gabriel Bucur
}

