% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_mi}
\alias{add_mi}
\title{Add Meta Information to R Environments}
\usage{
add_mi(series, map_list, meta_env = "meta_localized", overwrite = T)
}
\arguments{
\item{series}{character name key of}

\item{map_list}{list to represent key value mapping. Could also be of class miro.}

\item{meta_env}{name of the environment that holds the meta
environment. Defaults to meta_localized.}

\item{overwrite}{logical should existing meta information be overwritten inside
the environment?}
}
\description{
This function adds meta information to environments that
are explicitly meant to store Meta Information. This function
can be used separately in interactive R Session or to facilitate
mapping database information to R.
}

