% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsSetFunctions.R
\name{addKeysToTsSet}
\alias{addKeysToTsSet}
\title{Add keys to an existing Time Series set}
\usage{
addKeysToTsSet(con, set_name, ts_keys, user_name = Sys.info()["user"],
  tbl = "timeseries_sets", schema = "timeseries")
}
\arguments{
\item{con}{PostgreSQL connection}

\item{set_name}{The name of the set}

\item{ts_keys}{A character vector of keys to be added}

\item{user_name}{The user name of the set's owner}

\item{tbl}{Name of the time series sets table}

\item{schema}{Schema of the time series database to use}
}
\description{
Add keys to an existing Time Series set
}
\author{
Severin Thöni
}
