% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTables.R
\name{addReleaseDateToTimeseriesMain}
\alias{addReleaseDateToTimeseriesMain}
\title{Title}
\usage{
addReleaseDateToTimeseriesMain(schema = "timeseries",
  tbl = "timeseries_main")
}
\arguments{
\item{schema}{}

\item{tbl}{}
}
\description{
Title
}
