% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTables.R
\name{addReleaseDateToTimeseriesVintages}
\alias{addReleaseDateToTimeseriesVintages}
\title{Title}
\usage{
addReleaseDateToTimeseriesVintages(schema = "timeseries",
  tbl = "timeseries_vintages")
}
\arguments{
\item{schema}{}

\item{tbl}{}
}
\description{
Title
}
