% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{as.meta}
\alias{as.meta}
\title{Convert a List into a Metadata Object}
\usage{
as.meta(x)
}
\arguments{
\item{x}{list of meta data.}
}
\description{
Create timeseriesdb specific metadata class. Typically one list per natural language
is converted to a meta description object.
}
