% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{as.tsmeta}
\alias{as.tsmeta}
\title{Convert a List into a Metadata Object}
\usage{
as.tsmeta(meta, ...)
}
\arguments{
\item{meta}{list containing meta information. List elements are character strings.}

\item{...}{additional arguments, passed on to metthods below.}
}
\description{
Create timeseriesdb specific metadata class. Typically one list per natural language
is converted to a meta description object.
}
