% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_levels.R
\name{change_access_level}
\alias{change_access_level}
\alias{db_ts_change_access}
\alias{db_dataset_change_access}
\title{Change the Access Level of a Time Series}
\usage{
db_ts_change_access(
  con,
  ts_keys,
  access_level,
  valid_from = NULL,
  schema = "timeseries"
)

db_dataset_change_access(
  con,
  dataset,
  access_level,
  valid_from = NULL,
  schema = "timeseries"
)
}
\arguments{
\item{con}{RPostgres connection object.}

\item{ts_keys}{\strong{character} vector of time series identifiers.}

\item{access_level}{\strong{character} describing the access level of the time series or dataset.}

\item{valid_from}{character representation of a date in the form of 'YYYY-MM-DD'. valid_from starts a new version}

\item{schema}{\strong{character} name of the database schema. Defaults to 'timeseries'}

\item{dataset}{\strong{character} name of the dataset. Datasets are group of time series.}
}
\value{
returns a list containing the parsed JSON status feedback from the DB.

returns a list containing the parsed JSON status feedback from the DB.


}
\description{
Change the Access Level of a Time Series
}
\seealso{
Other access levels functions: 
\code{\link{db_access_level_create}()},
\code{\link{db_access_level_delete}()},
\code{\link{db_access_level_list}()},
\code{\link{db_access_level_set_default}()},
\code{\link{db_ts_find_keys}()}
}
\concept{access levels functions}
