% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsSetFunctions.R
\name{changeTsSetOwner}
\alias{changeTsSetOwner}
\title{Change the owner of a Time Series set}
\usage{
changeTsSetOwner(con, set_name, old_owner = Sys.info()["user"],
  new_owner, tbl = "timeseries_sets", schema = "timeseries")
}
\arguments{
\item{con}{PostgreSQL connection}

\item{set_name}{Name of the set to be updates}

\item{old_owner}{User name of the set's current owner}

\item{new_owner}{User name of the set's new owner}

\item{tbl}{Name of the time series sets table}

\item{schema}{Schema of the time series database to use}
}
\description{
Change the owner of a Time Series set
}
\author{
Severin Thöni
}
