% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{create_meta}
\alias{create_meta}
\title{Create Meta Data Objects}
\usage{
create_meta(...)
}
\arguments{
\item{...}{arguments passed on the respective method.}
}
\description{
Create list based S3 objects to store meta data. Meta data objects can 
be passed on to a plethora of functions including storing to database.
}
