% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createHstore}
\alias{createHstore}
\alias{createHstore.ts}
\title{Create Hstore}
\usage{
createHstore(x)

\method{createHstore}{ts}(x)
}
\arguments{
\item{x}{a time series object}
}
\description{
Function to Create Hstore Key Value Pair Mapping
}
\details{
This function creates a key value pair mapping from a time series object.
It returns an hstore object that can be inserted to a PostgreSQL database relation field of type hstore.
}
\examples{
ts1 <- ts(rnorm(100),start = c(1990,1),frequency = 4)
createHstore(ts1)
}
\author{
Matthias Bannert
}

