% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbExplorerUtils.R
\name{createUI}
\alias{createUI}
\alias{createUI.key}
\alias{createUI.set}
\alias{createUI.md}
\title{Polymorphic Helpers for the Shiny Based GUI}
\usage{
createUI(x, con, ...)

\method{createUI}{key}(x, con, ...)

\method{createUI}{set}(x, con, ...)

\method{createUI}{md}(x, con, ...)
}
\arguments{
\item{x}{character query type.}

\item{con}{PostgreSQL Connection object.}

\item{...}{additional arguments to be passed to the respective methods.}
}
\description{
Though R is not strictly OO it is sometimes useful to add some object orientation. 
In the case of a shiny app with an dynamically created user interface a bit
of pseudo polymorphism helps to avoid too many but,ifs and elses. 
Depending on the query type that is selected by the user a UI creation
method is called.
}
