% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_to_index}
\alias{date_to_index}
\title{Convert date-likes to time index}
\usage{
date_to_index(x)
}
\arguments{
\item{x}{The Date or Y-m-d string to convert}
}
\value{
The numeric representation of the date that can be used with ts
}
\description{
Convert date-likes to time index
}
\examples{
\dontrun{date_to_index("2020-07-01")}
}
