% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_levels.R
\name{db_access_level_set_default}
\alias{db_access_level_set_default}
\title{Set the Default Access Level}
\usage{
db_access_level_set_default(con, access_level, schema = "timeseries")
}
\arguments{
\item{con}{RPostgres connection object.}

\item{access_level}{\strong{character} describing the access level of the time series or dataset.}

\item{schema}{\strong{character} name of the database schema. Defaults to 'timeseries'}
}
\value{
returns a list containing the parsed JSON status feedback from the DB.
}
\description{
Changes the default access level. Apparently only one access level can be
the default level at a time.
}
\seealso{
Other access levels functions: 
\code{\link{change_access_level}},
\code{\link{db_access_level_create}()},
\code{\link{db_access_level_delete}()},
\code{\link{db_access_level_list}()},
\code{\link{db_ts_find_keys}()}
}
\concept{access levels functions}
