% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_connection_close}
\alias{db_connection_close}
\title{Close an Existing Database Connection}
\usage{
db_connection_close(con, ...)
}
\arguments{
\item{con}{RPostgres connection object.}

\item{...}{passed on to RPostgres::dbDisconnect}
}
\description{
Close database connection given a connection object.
}
