% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{db_dataset_read_metadata}
\alias{db_dataset_read_metadata}
\title{Read Dataset Meta Information}
\usage{
db_dataset_read_metadata(
  con,
  dataset_id,
  valid_on = NULL,
  locale = NULL,
  schema = "timeseries"
)
}
\arguments{
\item{con}{RPostgres connection object.}

\item{dataset_id}{character name of the dataset.}

\item{valid_on}{\strong{character} representation of a date in the form of 'YYYY-MM-DD'. valid_on selects the
version of a time series that is valid at the specified time.}

\item{locale}{character ISO-2 country locale.}

\item{schema}{\strong{character} name of the database schema. Defaults to 'timeseries'}
}
\description{
Read Dataset Meta Information
}
\seealso{
Other metadata functions: 
\code{\link{db_collection_read_metadata}()},
\code{\link{db_meta_get_latest_validity}()},
\code{\link{db_metadata_read}()},
\code{\link{db_metadata_store}()}
}
\concept{metadata functions}
