% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_get_installed_version}
\alias{db_get_installed_version}
\title{Get the Currently Installed Version of Timeseriesdb}
\usage{
db_get_installed_version(con, schema = "timeseries")
}
\arguments{
\item{con}{RPostgres connection object.}

\item{schema}{\strong{character} name of the database schema. Defaults to 'timeseries'}
}
\value{
character The version number of timeseriesdb currently installed on the given schema
}
\description{
Get the Currently Installed Version of Timeseriesdb
}
