% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_grant_to_admin}
\alias{db_grant_to_admin}
\title{GRANT all rights on a (temp) table to schema admin}
\usage{
db_grant_to_admin(con, table, schema = "timeseries")
}
\arguments{
\item{con}{RPostgres connection object.}

\item{table}{which table to grant rights on}

\item{schema}{\strong{character} name of the database schema. Defaults to 'timeseries'}
}
\description{
The SECURITY DEFINER functions do not have access to tables that
are stored via dbWriteTable. Usage rights on these tables must
be granted for them to be usable inside the db functions
}
