% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{db_release_cancel}
\alias{db_release_cancel}
\title{Cancel a Scheduled Release}
\usage{
db_release_cancel(con, release_id, schema = "timeseries")
}
\arguments{
\item{con}{RPostgres connection object.}

\item{release_id}{character ID of the release to cancel}

\item{schema}{\strong{character} name of the database schema. Defaults to 'timeseries'}
}
\description{
Attempts to cancel a release that has already passed will result in an error.
}
\seealso{
Other calendar functions: 
\code{\link{db_dataset_get_latest_release}()},
\code{\link{db_dataset_get_next_release}()},
\code{\link{db_dataset_get_release}()},
\code{\link{db_release_create}()},
\code{\link{db_release_list}()},
\code{\link{db_release_update}()}
}
\concept{calendar functions}
