% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{db_release_list}
\alias{db_release_list}
\title{List Data on Registered Releases}
\usage{
db_release_list(con, include_past = FALSE, schema = "timeseries")
}
\arguments{
\item{con}{RPostgres connection object.}

\item{include_past}{Should past releases be included? Defaults to FALSE}

\item{schema}{\strong{character} name of the database schema. Defaults to 'timeseries'}
}
\value{
data.frame with columns `id`, `title`, `note`, `release_date`, `reference_year`, `reference_period`, `reference_frequency`
}
\description{
List Data on Registered Releases
}
\seealso{
Other calendar functions: 
\code{\link{db_dataset_get_latest_release}()},
\code{\link{db_dataset_get_next_release}()},
\code{\link{db_dataset_get_release}()},
\code{\link{db_release_cancel}()},
\code{\link{db_release_create}()},
\code{\link{db_release_update}()}
}
\concept{calendar functions}
