% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_levels.R
\name{db_ts_get_access_level}
\alias{db_ts_get_access_level}
\title{Find Out About the Access Level of a Vintage}
\usage{
db_ts_get_access_level(con, ts_keys, valid_on = NULL, schema = "timeseries")
}
\arguments{
\item{con}{RPostgres connection object.}

\item{ts_keys}{\strong{character} vector of time series identifiers.}

\item{valid_on}{\strong{character} representation of a date in the form of 'YYYY-MM-DD'. valid_on selects the
version of a time series that is valid at the specified time.}

\item{schema}{\strong{character} name of the database schema. Defaults to 'timeseries'}
}
\description{
Provide the function with vector of time series keys and find out which access level is necessary to access the supplied keys.
}
