% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{db_ts_rename}
\alias{db_ts_rename}
\title{Rename Time Series by Assigning a New Key}
\usage{
db_ts_rename(con, ts_key, ts_key_new, schema = "timeseries")
}
\arguments{
\item{con}{RPostgres connection object.}

\item{ts_key}{character Vector of keys to rename}

\item{ts_key_new}{character Vector of new names}

\item{schema}{\strong{character} name of the database schema. Defaults to 'timeseries'}
}
\description{
Rename Time Series by Assigning a New Key
}
