% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tmp_tables.R
\name{db_with_tmp_read}
\alias{db_with_tmp_read}
\title{Helper to Create and Populate a Temporary Table for Fast Reading}
\usage{
db_with_tmp_read(con, ts_keys, regex = FALSE, code, schema = "timeseries")
}
\arguments{
\item{con}{RPostgres connection object.}

\item{ts_keys}{\strong{character} vector of time series identifiers.}

\item{regex}{logical if set to TRUE, the ts_keys parameter is interpreted as a regular expression pattern.}

\item{code}{expression Code to be evaluated after populating the temporary table on the database
of a time series that is valid from the specified date.}

\item{schema}{\strong{character} name of the database schema. Defaults to 'timeseries'}
}
\description{
This function is not exported. It creates a tempory table containing the
keys that should be read to join them against the time series storage.
This is much faster for larger selections than simple where clauses.
}
