% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{dbIsValid,PostgreSQLConnection-method}
\alias{dbIsValid,PostgreSQLConnection-method}
\alias{dbIsValid}
\title{Check Validity of a PostgreSQL connection}
\usage{
\S4method{dbIsValid}{PostgreSQLConnection}(dbObj)
}
\arguments{
\item{dbObj}{PostgreSQL connection object.}
}
\description{
Is the PostgreSQL connection expired?
}
