% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{dbIsValid}
\alias{dbIsValid}
\title{Check whether an PostgreSQL object is valid or not.}
\usage{
\S4method{dbIsValid}{PostgreSQLConnection}(dbObj)
}
\arguments{
\item{dbObj}{a PostgreSQLConnection connection object. Check if the connection has expired.}
}
\value{
A logical scalar.
}
\description{
Support function that verifies that the holding a reference to a
foreign object is still valid for communicating with the RDBMS
}

