% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsSetFunctions.R
\name{deactivateTsSet}
\alias{deactivateTsSet}
\title{Deactivate a Set of Time Series}
\usage{
deactivateTsSet(con, set_name, user_name = Sys.info()["user"],
  tbl = "timeseries_sets", schema = "timeseries")
}
\arguments{
\item{con}{PostgreSQL connection object}

\item{set_name}{character name of the set to be deactivated.}

\item{user_name}{character name of the user. Defaults to system user.}

\item{tbl}{character name of set tqble. Defaults to timeseries\_sets.}

\item{schema}{character name of the database schema. Defaults to timeseries.}
}
\description{
This deactivates a set of time series to get out of the user's sight, 
but it's not the deleted because users may not delete sets.
}
\author{
Matthias Bannert, Ioan Gabriel Bucur
}
