% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsSetFunctions.R
\name{deleteTsSet}
\alias{deleteTsSet}
\title{Permanently delete a Set of Time Series Keys}
\usage{
deleteTsSet(con, set_name, user_name = Sys.info()["user"],
  tbl = "timeseries_sets", schema = "timeseries")
}
\arguments{
\item{con}{PostgreSQL connection object}

\item{set_name}{The name of the set to be deleted}

\item{user_name}{Username to which the set belongs}

\item{tbl}{Name of set table}

\item{schema}{Name of timeseries schema}
}
\description{
Permanently delete a Set of Time Series Keys
}
\author{
Severin Thöni
}
