% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{depth}
\alias{depth}
\title{Determine depth of a list}
\usage{
depth(this)
}
\arguments{
\item{this}{an object of class list}
}
\description{
This function recursively checks the depth of a list and returns an integer value of depth
}
\details{
Hat tip to flodel at stackoverflow for suggesting this light weight way analyze depth of a nested list. Further complexity needs to be added to cover the fact that data.frame are lists, too. A more sophisticated recursive function can be found in the gatveys2 package.
}
\references{
http://stackoverflow.com/questions/13432863/determine-level-of-nesting-in-r
}

