% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exportMetaEnv.R
\name{exportMetaEnv}
\alias{exportMetaEnv}
\title{Export Content of Meta Information Environment to Various File Formats}
\usage{
exportMetaEnv(meta_env, fname = NULL, export_type = "pdf",
  flexcols = NULL, row.names = F, sep = ";", overwrite = T)
}
\arguments{
\item{meta_env}{meta\_env environment object.}

\item{fname}{character file name including file extension. If set to NULL no file
is export. The resulting data.frame is just displayed on the console in this case. Defaults to NULL.}

\item{export_type}{character indication which file format should be exported.
"pdf","tex","csv" are the eligible.}

\item{flexcols}{which columns shoukd be kept in the data.frame. Defaults to NULL, using all columns.}

\item{row.names}{logical should row.names be displayed in csv.}

\item{sep}{character seperator}

\item{overwrite}{should existing files be overwritten? Defaults to TRUE.}
}
\description{
The idea of this function is to create a standalone meta information catalog.
The catalog file can be used as a companion to illustrate time series exports from
timeseriesdb. Note that this function imports functionality from other packages such as
data.table and openxlsx.
}

