% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fetchSelectedResults}
\alias{fetchSelectedResults}
\title{Fetch Selected Results from Database}
\usage{
fetchSelectedResults(result_set, connect = "con", envir = NULL,
  tbl = "timeseries_main")
}
\arguments{
\item{result_set}{object of class rs}

\item{connect}{character name of the PostgreSQL connection object, defaults to 'con'.}

\item{tbl}{character variable indicating the name the timeseries table in the databse.}

\item{envir}{environment to assign fetched time series to. If not specified a list of time series is returned. Defaults to NULL.
Defaults to timeseries_main.
'con' in the global environment.}
}
\description{
This function uses result sets selections to read multiple
time series from a database.
}
\author{
Matthias Bannert
}

