% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMeta}
\alias{getMeta}
\title{Quickly Handle Meta Information}
\usage{
getMeta(series, lang, con = options()$TIMESERIESDB_CON,
  tbl = "meta_data_localized")
}
\arguments{
\item{series}{an R time series object}

\item{lang}{character name of the language of the meta information. Typically 'de', 'it', 'fr' or 'en'.}

\item{con}{connection object, defaults to object set by options().}

\item{tbl}{character name of the table that contains the meta information.}
}
\description{
Sometimes reading the entire meta description for all language or
multiple time series might not be necessary. Quick handle operators
help users to access the information quickly as a non-nested list
for only one language is returned. These functions are alpha status,
more will follow.
}

