% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeSeriesVintages.R
\name{getTimeSeriesVintages}
\alias{getTimeSeriesVintages}
\title{Get all available vintages for the time series identified by \code{series}}
\usage{
getTimeSeriesVintages(series, con, tbl_vintages = "timeseries_vintages",
  schema = "timeseries")
}
\arguments{
\item{series}{character Names of the time series for which to get the vintages}

\item{con}{PostgreSQL connection object.}

\item{tbl_vintages}{character string denoting the name of the vintages time series table in the PostgreSQL database.}

\item{schema}{SQL schema name. Defaults to timeseries.}
}
\description{
Get all available vintages for the time series identified by \code{series}
}
