% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imortTsList.R
\name{importTsList}
\alias{importTsList}
\title{Import A Long Format CSV File Into a List of Time Series}
\usage{
importTsList(f, base_ts = T)
}
\arguments{
\item{f}{character file name}

\item{base_ts}{logical should the export be casted to R basic time series or left as xts? Defaults to TRUE.}
}
\description{
Generates a List of time series of class tslist from a flat file that contains multiple time series.
}

