% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_to_ts.R
\name{json_to_ts}
\alias{json_to_ts}
\title{Convert JSON Representation of a Time Series into R Time Series Objects}
\usage{
json_to_ts(jsn, as.dt = FALSE)
}
\arguments{
\item{jsn}{JSON string to convert}

\item{as.dt}{boolean Should the result be returned as a data.table?}
}
\value{
R time series representation of class ts, xts or data.table depending on parameter setting and nature of time series. Regular time series can be returned as 'ts' objects whereas irregular time series use 'xts' objects.
}
\description{
This function is not exported.
}
