% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kof_ts}
\alias{kof_ts}
\title{KOF indicators}
\format{
A list with four time series objects:
\describe{
 \item{ch.kof.barometer}{Indicator for the Swiss Business Cycle.}
 \item{baro}{Vintages (versions) of the KOF Barometer Indicator.}
 \item{ch.kof.ie.retro.ch_total.ind.d11}{KOF Employment Indicator for Switzerland}
}
}
\source{
KOF Swiss Economic Institute - KOF indicators.
\url{https://kof.ethz.ch/en/forecasts-and-indicators/indicators.html}
}
\usage{
kof_ts
}
\description{
KOF indicators
}
\examples{
\dontrun{
 kof_ts
}
}
\keyword{datasets}
