% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tsSetFunctions.R
\name{listTsSets}
\alias{listTsSets}
\title{List All Time Series Sets for a Particular User}
\usage{
listTsSets(con, user_name = Sys.info()["user"], tbl = "timeseries_sets",
  schema = "timeseries")
}
\arguments{
\item{con}{PostgreSQL connection object}

\item{user_name}{character name of the user. Defaults to system user.}

\item{tbl}{character name of set tqble. Defaults to timeseries\_sets.}

\item{schema}{character name of the database schema. Defaults to timeseries.}
}
\description{
Show the names of all sets that are available to a particular user.
}
\author{
Matthias Bannert, Gabriel Bucur
}

