% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgCopyDf.R
\name{pgCopyDf}
\alias{pgCopyDf}
\title{Copy data.frame to postgres using bulk copy}
\usage{
pgCopyDf(con, d, q, chunksize = 10000)
}
\arguments{
\item{con}{PostgreSQL connection object.}

\item{d}{data.frame}

\item{q}{}

\item{chunksize}{integer, defaults to 10000.}
}
\description{
Copy data.frame to postgres using bulk copy
}
