% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotMethod.R
\name{plot.tslist}
\alias{plot.tslist}
\title{Plot a list of Time Series
Directly plot time series of mixed frequency into one plot.
\code{\link{readTimeSeries}} returns a list of time series. The returned list is
also of class tslist which makes it very convenient to plot timeseries,
directly out of the database. This function gets automatically the right ranges
for the axes to plot all values and dates in contained in the list.}
\usage{
\method{plot}{tslist}(x, ...)
}
\arguments{
\item{x}{a list of time series. Object should be of class tslist.}

\item{...}{parameters than can simply by passed on tot the plot function}
}
\description{
Plot a list of Time Series
Directly plot time series of mixed frequency into one plot.
\code{\link{readTimeSeries}} returns a list of time series. The returned list is
also of class tslist which makes it very convenient to plot timeseries,
directly out of the database. This function gets automatically the right ranges
for the axes to plot all values and dates in contained in the list.
}

