% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{print.meta}
\alias{print.meta}
\title{Print Method for meta Object}
\usage{
\method{print}{meta}(x, ...)
}
\arguments{
\item{x}{a metadata object.}

\item{...}{list of print options.}
}
\description{
Print Method for meta Object
}
