% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%mdb\%}
\alias{\%mdb\%}
\title{Quick Handle operators}
\usage{
series \%mdb\% lang
}
\arguments{
\item{series}{an R time series object}

\item{lang}{character representation of a language, typically 'en', 'de', 'fr' or 'it' as this is a Swiss package.}

\item{con}{a Postgres Connection object, typically set in in options().}

\item{tbl}{character name of the table that contains localized meta information}
}
\description{
Sometimes reading the entire meta description for all language or
multiple time series might not be necessary. Quick handle operators
help users to access the information quickly as a non-nested list
for only one language is returned. These operators are alpha status,
more will follow.
}

