% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAllVintages.R
\name{readAllVintages}
\alias{readAllVintages}
\title{Read All Vintages that Correspond to a Particular Time Series Key}
\usage{
readAllVintages(con, series, tbl_vintages = "timeseries_vintages",
  schema = "timeseries", pkg_for_irreg = "xts",
  respect_release_date = FALSE)
}
\arguments{
\item{con}{PostgreSQL connection object}

\item{series}{character identiffier of the time series (ts_key).}

\item{tbl_vintages}{character name of the vintages table. 
Defaults to "timeseries_vintages"}

\item{schema}{character name of the database schema. Defaults to "timeseries".}

\item{pkg_for_irreg}{character name of the package used to handle irregular
time series. Defaults to "xts".}

\item{respect_release_date}{logical Should release dates be respected?
Defaults to FALSE. Typically only used by APIs for external users.}
}
\description{
Data in official statistics often get revised after being published. Hence it
is imported for (academic) research to keep track of all vintages of a time series
because the most recent version of a time series on an instution's website etc. might
not correspond to what was used in an older research project. For the sake of 
reproducibility timeseriesdb offers the opportunity to store time series 
vintages and makes sure time series do not overlap and older vintages are not
overwritten. This function reads all vintages for one single key and returns
a list of time series vintages.
}
