% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readMetaInformation}
\alias{readMetaInformation}
\title{Read Meta Information from a Time Series Database}
\usage{
readMetaInformation(series, con = options()$TIMESERIESDB_CON, overwrite,
  localized = T, tbl = "meta_data_localized")
}
\arguments{
\item{series}{character name of a time series object.}

\item{con}{PostgreSQL connection object}

\item{overwrite}{logical should data be overwritten}

\item{localized}{logical should localized meta information be read, defaults to TRUE.}

\item{tbl}{character name of the table that contains the
localized meta information}
}
\description{
This function reads meta information from timeseriesdb package PostgreSQL
database and puts into a meta information environment.
}

