% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readTimeseries}
\alias{readTimeseries}
\title{Read Time Series From PostgreSQL database}
\usage{
readTimeseries(series, connect = "con", tbl = "timeseries_main")
}
\arguments{
\item{series}{character representation of the key of the time series}

\item{connect}{character name of the PostgreSQL connection object}

\item{tbl}{character string denoting the name of the main time series table in the PostgreSQL database.}
}
\description{
This function reads a time series from a postgres database,
that key value pair storage (hstore), to archive time series.
After reading the information from the database a standard
R time series object of class 'ts' is built and returned.
}
\author{
Matthias Bannert
}

