% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsSetFunctions.R
\name{removeKeysFromTsSet}
\alias{removeKeysFromTsSet}
\title{Remove keys from a Time Series set (if present)}
\usage{
removeKeysFromTsSet(con, set_name, ts_keys,
  user_name = Sys.info()["user"], tbl = "timeseries_sets",
  schema = "timeseries")
}
\arguments{
\item{con}{PostgreSQL connection}

\item{set_name}{character name of a time series set.}

\item{ts_keys}{A character vector of keys to be removed.}

\item{user_name}{The user name of the set's owner.}

\item{tbl}{Name of the time series sets table.}

\item{schema}{Schema of the time series database to use.}
}
\description{
Remove keys from a Time Series set (if present)
}
\author{
Severin Thöni
}
