% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resolveOverlap}
\alias{resolveOverlap}
\title{Concatenate Time Series and Resolve Overlap Automatically}
\usage{
resolveOverlap(ts1, ts2)
}
\arguments{
\item{ts1}{ts time series, typically the older series}

\item{ts2}{ts time series, typically the younger series}
}
\description{
Resolve overlap determines which of two ts class time series is
reaching further and arranges the two series into first and second
series accordingly. Both time series are concatenated to one
if both series had the same frequency.
}

