% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmAllBut}
\alias{rmAllBut}
\title{Delete all objects except for specific objects}
\usage{
rmAllBut(but, env = .GlobalEnv, quiet = F)
}
\arguments{
\item{but}{character vector of variables that should not be deleted.}

\item{env}{environment to clean up. Defaults to .Globalenv}

\item{quiet}{logical should functions print output? Defaults to falase.}
}
\description{
Run rm(list=ls()) but sparing some objects from being deleted. 
This function is particularly handy when you want to clear the memory but want to keep the 
the database connection object.
}

