% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rollback}
\alias{rollback}
\title{Rollback all Deleted Time Series to Database}
\usage{
rollback(rollback_env = "db_rollback", connect = "con",
  tbl = "timeseries_main")
}
\arguments{
\item{rollback_env}{character name of rollback environment.}

\item{connect}{character name of the PostgreSQL connection object.}

\item{tbl}{character variable name of the time series table in the database. Defaults to 'timeseries_main'.}
}
\description{
This function rolls back all deleted time series that were deleted by \code{\link{deleteTimeseries}} and thus were stored in rollback environment in memory. This function only works if the option rollback = TRUE was set when exectuing the delete query.
}
\author{
Matthias Bannert
}

