% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTables.R
\name{runCreateTables}
\alias{runCreateTables}
\title{Run Setup: Create all mandatory tables}
\usage{
runCreateTables(con, schema = "timeseries")
}
\arguments{
\item{con}{PostgreSQL connection object. Typically created with \code{\link{createConObj}}.}

\item{schema}{character denoting a PostgreSQL schema.}
}
\description{
Creates all tables absolutely needed for timeseriesdb to work correctly. 
This function should only be run once as an initial setup. Make sure you got sufficient 
access rights. The function returns a list of status reports for the its 5 database queries. 
look at this helps you to see whether anything went wrong.
}

