% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDbQuery.R
\name{runDbQuery}
\alias{runDbQuery}
\title{Run database query}
\usage{
runDbQuery(con, sql_query)
}
\arguments{
\item{con}{PostgreSQL connection object}

\item{sql_query}{character string containing a SQL query}
}
\description{
Run database queries using \code{\link{DBI::dbSendQuery}}, \code{\link{DBI::fetch}} and \code{\link{DBI::dbClearResult}} in similar fashion as \code{\link{DBI::dbGetQuery}} but provide better error handling. 
This function always returns a data.frame as opposed to different types in case of an exception. However, if the database query fails and empty data.frame is returned. Besides query status and database error are returned as attributes.
}
\examples{
# There's no connection, so this returns a proper error message.
out_obj <- runDbQuery(bogus_connection,"SELECT * FROM some_table") 
attributes(out_obj)
}

