% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTables.R
\name{runUpgradeTables}
\alias{runUpgradeTables}
\title{Add Release Date Column to Tables}
\usage{
runUpgradeTables(con, schema = "timeseries")
}
\arguments{
\item{con}{PostgreSQLL connection object}

\item{schema}{database schema, defaults to 'timeseries'.}
}
\description{
Adds a release column to tables of older versions of timeseriesdb.
}
