% Generated by roxygen2 (4.0.2): do not edit by hand
\name{searchDbForKeys}
\alias{searchDbForKeys}
\title{Search Database for keys based on string wildcards}
\usage{
searchDbForKeys(..., connect = "con", tbl = "timeseries_main")
}
\arguments{
\item{...}{character string patterns.}

\item{connect}{character name of the PostgreSQL connection object.}

\item{tbl}{character representation of the name of the main timeseries
table. Defaults to 'timeseries_main'}
}
\description{
This function searches for keys that match a particular
chunk of the key and returns all matches in a result set.
}
\author{
Matthias Bannert
}

