% Generated by roxygen2 (4.0.2): do not edit by hand
\name{searchKVP}
\alias{searchKVP}
\title{Search Key-Value Pairs}
\usage{
searchKVP(key, value, con = options()$TIMESERIESDB_CON,
  hstore = "meta_data", tbl = "meta_data_unlocalized")
}
\arguments{
\item{key}{character}

\item{value}{in the hstore}

\item{con}{PostgreSQL connection object}

\item{hstore}{name of the hstore column}

\item{tbl}{name of the table to be queried. defaults to  'meta_data_localized'}
}
\description{
Search hstore key value in PostgreSQL. Very handsome when crawling the database by meta informaiton.
Currently works for non translated meta information.
}

