% Generated by roxygen2 (4.0.2): do not edit by hand
\name{selectResults}
\alias{selectResults}
\title{Select Results in a Result Set for Fetching from Database}
\usage{
selectResults(result_set, selection)
}
\arguments{
\item{result_set}{object of class rs}

\item{selection}{numeric vector indicating line numbers of selected time series.}
}
\description{
This function lets the user select multiple time series from
a result set that can be fetched from the database subsequently.
To do so use \code{\link{fetchSelectedResults}}.
}
\author{
Matthias Bannert
}

