% Generated by roxygen2 (4.0.2): do not edit by hand
\name{serveDublinCore}
\alias{serveDublinCore}
\title{Serve Dublin Core Compliant Meta Information for a Single Record}
\usage{
serveDublinCore(id, lang = "de", lookup_localized = "meta_data_localized",
  lookup_unlocalized = "meta_data_unlocalized", contibutor = NULL,
  creator = NULL, description = NULL, format = NULL, language = NULL,
  publisher = NULL, rights = NULL, src = NULL, subject = "",
  title = NULL)
}
\arguments{
\item{id}{character identifier, time series key.}

\item{lookup_localized}{}
}
\description{
This function is experimental. It is meant to serve Dublin Core compliant meta information for single time series records.
}

