% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sql.R
\name{setup_sql_extentions}
\alias{setup_sql_extentions}
\title{Install PostgreSQL Schemas and Extensions}
\usage{
setup_sql_extentions(con, schema = "timeseries")
}
\arguments{
\item{con}{RPostgres connection object.}

\item{schema}{schema character schema name, defaults to 'timeseries'.}
}
\description{
Installs schema, uuid-ossp, btree_gist.
This function must be run with a connection of a database level admin.
}
