% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sql.R
\name{setup_sql_grant_rights}
\alias{setup_sql_grant_rights}
\title{Grant execute on {timeseriesdb} functions}
\usage{
setup_sql_grant_rights(con, schema = "timeseries", prnt = identity)
}
\arguments{
\item{con}{RPostgres connection object}

\item{schema}{character schema name, defaults to 'timeseries'}

\item{prnt}{function log printing function}
}
\description{
Grant execute on {timeseriesdb} functions
}
