% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sql.R
\name{setup_sql_roles}
\alias{setup_sql_roles}
\title{Create Roles needed for operation of timeseriesdb}
\usage{
setup_sql_roles(con, schema = "timeseries")
}
\arguments{
\item{con}{RPostgres connection object}

\item{schema}{schema character schema name, defaults to 'timeseries'.}
}
\description{
This function must be run with a connection of a database level admin.
}
