% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sql.R
\name{setup_sql_tables}
\alias{setup_sql_tables}
\title{Install {timeseriesdb} System Tables}
\usage{
setup_sql_tables(con, schema = "timeseries", prnt = identity)
}
\arguments{
\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{schema}{character schema name, defaults to 'timeseries'.}

\item{prnt}{function log printing function}
}
\description{
Installs tables needed to operated {timeseriesdb} in
a given PostgreSQL schema. The tables use a default SQL file installed
with the package to generate SQL tables. The default schema 'timeseries'
can be replaced using the 'schema' parameter.
}
