% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sql.R
\name{setup_sql_triggers}
\alias{setup_sql_triggers}
\title{Install {timeseriesdb} Triggers}
\usage{
setup_sql_triggers(con, schema = "timeseries", prnt = identity)
}
\arguments{
\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{schema}{character schema name, defaults to 'timeseries'.}

\item{prnt}{function log printing function}
}
\description{
Installs functions needed for timeseriesdb triggers and sets up these triggers in
a given PostgreSQL schema. The functions uses a default SQL file installed
with the package to generate SQL functions. The default schema 'timeseries'
can be replaced using the 'schema' parameter.
}
