% Generated by roxygen2 (4.0.2): do not edit by hand
\name{storeMetaInformation}
\alias{storeMetaInformation}
\title{Store Meta Information to the Database}
\usage{
storeMetaInformation(series, con = options()$TIMESERIESDB_CON,
  tbl = "meta_data_localized", lookup_env = "meta", overwrite = T)
}
\arguments{
\item{series}{a character name of an time series object}

\item{con}{a PostgreSQL connection object}

\item{tbl}{name of the meta information table, defaults to localized meta data: meta_data_localized. Alternatively choose meta_data_unlocalized if you are not translating meta information.}

\item{lookup_env}{name of the R environment in which to look for meta information objects}

\item{overwrite}{logical, TRUE}
}
\description{
This function stores meta information to the database for a given time series.
Make sure that corresponding time series had been inserted to the main table before.
}

