% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tsSetFunctions.R
\name{storeTsSet}
\alias{storeTsSet}
\title{Store a New Set of Time Series}
\usage{
storeTsSet(con, set_name, set_keys, user_name = Sys.info()["user"],
  description = "", active = TRUE, tbl = "timeseries_sets")
}
\arguments{
\item{con}{PostgreSQL connection object}

\item{set_name}{character name of a set time series in the database.}

\item{set_keys}{list of keys contained in the set and their type of key.}

\item{user_name}{character name of the user. Defaults to system user.}

\item{tbl}{character name of set tqble. Defaults to timeseries\_sets.}
}
\description{
Store a new set of Time Series to the database. Users can select the time series keys
that should be grouped inside a set.
}
\author{
Ioan Gabriel Bucur
}

