% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storeVintages.R
\name{storeVintages}
\alias{storeVintages}
\title{Write a vintages of time series to the database}
\usage{
storeVintages(series, con, vintage_range, li, store_freq = T,
  tbl = "vintages", md_unlocal = "vintages_meta_data", overwrite = T,
  schema = "timeseries")
}
\arguments{
\item{series}{character name of a time series, S3 class ts. When used with lists it is convenient to set series to names(li). Note that the series name needs to be unique in the database!}

\item{con}{a PostgreSQL connection object.}

\item{li}{list of time series. Defaults to NULL to no break legacy calls that use lookup environments.}

\item{tbl}{character string denoting the name of the main time series table in the PostgreSQL database.}

\item{md_unlocal}{character string denoting the name of the table that holds unlocalized meta information.}

\item{overwrite}{logical should existing records (same primary key) be overwritten? Defaults to TRUE.}

\item{schema}{SQL schema name. Defaults to timeseries.}
}
\description{
This function uses INSERT INTO instead of the more convenient dbWritetable for performance reasons. DO NOT USE THIS FUNCTIONS IN LOOPS OR LAPPLY! This function can handle a set of time series on its own and is much faster than looping over a list. Non-unique primary keys are overwritten !
}
\author{
Matthias Bannert
}

