% Generated by roxygen2 (4.0.2): do not edit by hand
\name{beginTransaction}
\alias{beginTransaction}
\title{Convenience Wrapper to SQL classics for BEGIN,COMMIT, ROLLBACK}
\usage{
beginTransaction(con = options()$TIMESERIESDB_CON)
}
\arguments{
\item{con}{PostgreSQL connection object.}
}
\description{
this set of function can speed up loops by starting a transaction,
performing several queries and ending them with either commit or rollback.
}

