% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/writeLogFile.R
\name{writeLogFile}
\alias{writeLogFile}
\title{Simple Log File Writer}
\usage{
writeLogFile(filename = NULL, msg, line_end = "\\n")
}
\arguments{
\item{filename}{character name of a textfile. Defaults to NULL.}

\item{msg}{log file message}

\item{line_end}{line end character}
}
\description{
Most simple log file writer just write steps of
a script to a text file.
}

