% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tsSetFunctions.R
\name{activateTsSet}
\alias{activateTsSet}
\title{Activate a Set of Time Series}
\usage{
activateTsSet(con, set_name, user_name = Sys.info()["user"],
  tbl = "timeseries_sets", schema = "timeseries")
}
\arguments{
\item{con}{PostgreSQL connection object}

\item{set_name}{character name of the set to be activated.}

\item{user_name}{character name of the user. Defaults to system user.}

\item{tbl}{character name of set tqble. Defaults to timeseries\_sets.}

\item{schema}{character name of the database schema. Defaults to timeseries.}
}
\description{
Activate a set of time series to get in the user's sight.
Deactivated sets are not deleted though.
}
\author{
Matthias Bannert, Ioan Gabriel Bucur
}

