% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addMetaInformation}
\alias{addMetaInformation}
\title{Add Meta Information to R Environments}
\usage{
addMetaInformation(series, map_list, meta_env = NULL, overwrite = T)
}
\arguments{
\item{series}{character name key of}

\item{map_list}{list to represent key value mapping. Could also be of class miro.}

\item{meta_env}{an environment that already holds meta information and should be extended.
Defaults to NULL in which case it creates and returns a new environment.}

\item{overwrite}{logical should existing meta information be overwritten inside
the environment?}
}
\description{
This function adds meta information to environments that
are explicitly meant to store Meta Information. This function
can be used separately in interactive R Session or to facilitate
mapping database information to R.
}

