% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dbExplorerUtils.R
\name{createChoices}
\alias{createChoices}
\alias{createChoices.key}
\alias{createChoices.md}
\alias{createChoices.set}
\title{Create Choices Boxes Dynamically}
\usage{
createChoices(x, input = NULL, keys = NULL, ...)

\method{createChoices}{key}(x, input = NULL, keys = NULL, ...)

\method{createChoices}{md}(x, input = NULL, keys = NULL, ...)

\method{createChoices}{set}(x, input = NULL, keys = NULL, ...)
}
\description{
Create Choices Boxes Dynamically
}

