% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createHstore}
\alias{createHstore}
\alias{createHstore.data.frame}
\alias{createHstore.list}
\alias{createHstore.miro}
\alias{createHstore.ts}
\title{Create Hstore}
\usage{
createHstore(x, ...)

\method{createHstore}{ts}(x, ...)

\method{createHstore}{data.frame}(x, ...)

\method{createHstore}{list}(x, ...)

\method{createHstore}{miro}(x, ...)
}
\arguments{
\item{x}{a time series object, a two column data frame or object of S3 class
miro (meta information for R objects).}

\item{...}{optional arguments, e.g. position of the key col and
pasition of the value col in a data.frame.}
}
\description{
Function to Create Hstore Key Value Pair Mapping
}
\details{
This function creates a key value pair mapping from a time series object.
It returns an hstore object that can be inserted to a PostgreSQL database relation field of type hstore.
}
\examples{
ts1 <- ts(rnorm(100),start = c(1990,1),frequency = 4)
createHstore(ts1)
}
\author{
Matthias Bannert
}

