% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHstore.R
\name{createHstore}
\alias{createHstore}
\alias{createHstore.ts}
\alias{createHstore.zoo}
\alias{createHstore.data.frame}
\alias{createHstore.list}
\title{Create Hstore}
\usage{
createHstore(x, ...)

\method{createHstore}{ts}(x, ...)

\method{createHstore}{zoo}(x, ...)

\method{createHstore}{data.frame}(x, key_col_index = 1)

\method{createHstore}{list}(x, ...)
}
\arguments{
\item{x}{a time series object, a two column data frame or object of S3 class
miro (meta information for R objects).}

\item{...}{optional arguments, fct = TRUE create text expressions of hstore function calls.
also for data.frames key_pos and value_pos could be given if they are different from 1 and 2. 
 e.g. position of the key col and
pasition of the value col in a data.frame.}
}
\description{
Function to Create Hstore Key Value Pair Mapping
}
\details{
This function creates a key value pair mapping from a time series object. 
It returns an hstore object that can be inserted to a PostgreSQL database relation field of type hstore.
}
\examples{
ts1 <- ts(rnorm(100),start = c(1990,1),frequency = 4)
createHstore(ts1)

}
\author{
Matthias Bannert
}
