% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dbExplorerUtils.R
\name{createUI}
\alias{createUI}
\alias{createUI.key}
\alias{createUI.md}
\alias{createUI.set}
\title{Polymorphic Helpers for the Shiny Based GUI}
\usage{
createUI(x, ...)

\method{createUI}{key}(x, ...)

\method{createUI}{set}(x, ...)

\method{createUI}{md}(x, ...)
}
\description{
Though R is not strictly OO it is sometimes useful to add some object orientation.
In the case of a shiny app with an dynamically created user interface a bit
of pseudo polymorphism helps to avoid too many but,ifs and elses.
Depending on the query type that is selected by the user a UI creation
method is called.
}

