% Generated by roxygen2 (4.0.1): do not edit by hand
\name{deleteTimeseries}
\alias{deleteTimeseries}
\title{Delete Time Series from Database}
\usage{
deleteTimeseries(series, connect = "con", rollback = T,
  tbl = "timeseries_main")
}
\arguments{
\item{series}{character name of the series that should be deleted}

\item{connect}{character name of the PostgreSQL connection object}

\item{rollback}{logical if TRUE the rollback functionality is enabled. Defaults
to TRUE.}

\item{tbl}{character name of the table that holds the timeseries in the database.
Defaults to 'timeseries_main'.}
}
\description{
This function deletes a time series from the database.
It also has a rollback feature which stores a series to an R environment and
keeps the deleted time series in memory as long as the R Session is alive or the
environment is manually deleted. Note that your database user needs to rights to
delete records from the database if this function should work.
}
\author{
Matthias Bannert
}

