% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreDb.R
\name{exploreDb}
\alias{exploreDb}
\title{Start a GUI to Explore Data}
\usage{
exploreDb(con, browser = T)
}
\arguments{
\item{con}{PostgreSQL Connection object}

\item{browser}{logical should app be fired up in the web browser? Defaults to TRUE.}
}
\description{
Start a graphical user interface in the user's standard web browser to search 
and explore time series data. Data can be searched using regular expressions for
keys. Hits can subsequently selected from a select box and are plotted in a joint
time series plot.
}

