% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exportTsList.R
\name{exportTsList}
\alias{exportTsList}
\title{Export A List of Time Series to CSV}
\usage{
exportTsList(tl, fname = NULL, cast = T, sep = ";", dec = ".")
}
\arguments{
\item{tl}{list of time series}

\item{fname}{character file name. If set to NULL a standard file name chunk + Sys.Date is used.}

\item{cast}{logical. Should the resulting data.frame be cast to wide format? Defaults to TRUE}

\item{sep}{character that separates columns in .csv}

\item{dec}{character that separates decimals in .csv}
}
\description{
Export a List of time series to semi-colon separated csv file. Typically multiple time series are exported as long format (melted format) files.
}

