% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exportTsList.R
\name{exportTsList}
\alias{exportTsList}
\title{Export A List of Time Series to CSV}
\usage{
exportTsList(tl, fname = NULL, cast = T, xlsx = F, sep = ";",
  dec = ".", LC_TIME_LOCALE = NULL, date_format = NULL)
}
\arguments{
\item{tl}{list of time series}

\item{fname}{character file name. If set to NULL a standard file name chunk + Sys.Date is used.}

\item{cast}{logical. Should the resulting data.frame be cast to wide format? Defaults to TRUE}

\item{xlsx}{logical. Should data be exported to .xlsx? Defaults to FALSE.}

\item{sep}{character that separates columns in .csv}

\item{dec}{character that separates decimals in .csv}

\item{LC_TIME_LOCALE}{character time locale that differs from the standard locale. e.g. en_US.UTF-8. Defaults to NULL and uses the standard locale then.}

\item{date_format}{character denotes the date format. Defaults to NULL. If set to null the default is used: Jan 2010. In combination with LC\_TIME\_Locale various international date formats can be produced.}
}
\description{
Export a List of time series to semi-colon separated csv file. Typically multiple time series are exported as long format (melted format) files.
}
\examples{
tslist <- list()
tslist$ts1 <- ts(rnorm(50),start = c(1990,1),frequency = 12)
Sys.getlocale() # gets all locale categories
Sys.getlocale("LC_TIME") # gets the time category only
exportTsList(tslist,LC_TIME_LOCALE = "de_DE.UTF-8",date_format="\%Y \%b")
# returns .csv file with a 2010-Mai style dates.
exportTsList(tslist)
# returns 2010-05 style dates
# Quarterly data does also work with format
tslist_q <- list()
tslist_q$ts2 <- ts(rnorm(50),start = c(1990,4),frequency = 4)
exportTsList(tslist_q,date_format="\%Y-0\%q")
}

