% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeta.R
\name{getMeta}
\alias{getMeta}
\title{Quickly Handle Meta Information}
\usage{
getMeta(series, lang, con, tbl = "meta_data_localized",
  schema = "timeseries")
}
\arguments{
\item{series}{an R time series object}

\item{lang}{character name of the language of the meta information. Typically 'de', 'it', 'fr' or 'en'.}

\item{con}{connection object}

\item{tbl}{character name of the table that contains the meta information.}

\item{schema}{SQL schema name. Defaults to 'timeseries'.}
}
\description{
Sometimes reading the entire meta description for all language or
multiple time series might not be necessary. Quick handle operators
help users to access the information quickly as a non-nested list 
for only one language is returned. These functions are alpha status, 
more will follow.
}
