% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/quickHandleOps.R
\name{\%k\%}
\alias{\%k\%}
\title{Search the Database by Keys}
\usage{
conObj \%k\% regexp
}
\arguments{
\item{conObj}{PostgreSQL Connection object.}

\item{regexp}{character regular expression pattern.}
}
\description{
Quick handle operator to search the database by keys. All time series whose key fit
the regular expression which was handed to the operator are returned in a list.
}

