% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readMetaInformation}
\alias{readMetaInformation}
\title{Read Meta Information from a Time Series Database}
\usage{
readMetaInformation(series, con = options()$TIMESERIESDB_CON, overwrite,
  locale = "de", tbl = "meta_data_localized")
}
\arguments{
\item{series}{character name of a time series object.}

\item{con}{PostgreSQL connection object}

\item{overwrite}{logical should data be overwritten}

\item{locale}{character denoting the locale of the meta information that is queried.
defaults to 'de' for German. At the KOF Swiss Economic Institute meta information should be available
als in English 'en', French 'fr' and Italian 'it'. Set the locale to NULL to query unlocalized meta information.}

\item{tbl}{character name of the table that contains meta information. Defaults to 'meta_data_localized'. If choose meta 'meta_data_unlocalized' when locale is set to NULL.}
}
\description{
This function reads meta information from a timeseriesdb package PostgreSQL
database and puts into a meta information environment.
}

