% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMetaInformation.R
\name{readMetaInformation}
\alias{readMetaInformation}
\title{Read Meta Information from a Time Series Database}
\usage{
readMetaInformation(series, con, locale = "de", tbl = "meta_data_localized",
  schema = "timeseries", as_list = TRUE)
}
\arguments{
\item{series}{character name of a time series object.}

\item{con}{PostgreSQL connection object}

\item{locale}{character denoting the locale of the meta information that is queried.
defaults to 'de' for German. At the KOF Swiss Economic Institute meta information should be available
als in English 'en', French 'fr' and Italian 'it'. Set the locale to NULL to query unlocalized meta information.}

\item{tbl}{character name of the table that contains meta information. Defaults to 'meta_data_localized'.
Choose meta 'meta_data_unlocalized' when locale is set to NULL.}

\item{schema}{SQL schema name. Defaults to timeseries.}

\item{as_list}{Should the result be returned as a tsmeta.list instead of a tsmeta.dt? Default TRUE}
}
\description{
This function reads meta information from a timeseriesdb package PostgreSQL
database and puts into a meta information environment.
}
