% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readMetaInformation}
\alias{readMetaInformation}
\title{Read Meta Information From the Database Given a Time Series Key}
\usage{
readMetaInformation(series, meta_fixed = TRUE, meta_localized = T,
  connect = "con", meta_table_fixed = "meta_data_unlocalized",
  meta_table_localized = "meta_data_localized")
}
\arguments{
\item{series}{character string time series key}

\item{meta_fixed}{logical should non-translatable data be queried? defaults to  to FALSE. Gets information such as legacy key or frequency.}

\item{meta_localized}{character either a two character language abbreviation or "all" which fetches all available meta information for a particular series.}

\item{connect}{character string name of a database connection object.}
}
\description{
This function is not really operational yet. WARNING
}
\author{
Matthias Bannert
}

