% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readTimeseries.R
\name{readTimeSeries}
\alias{readTimeSeries}
\title{Read Time Series From PostgreSQL database}
\usage{
readTimeSeries(series, con = Sys.getenv("TIMESERIESDB_CON"),
  tbl = "timeseries_main")
}
\arguments{
\item{series}{character vector of series names.}

\item{con}{a PostgreSQL connection object}

\item{tbl}{character string denoting the name of the main time series table
in the PostgreSQL database.}
}
\description{
This function reads a time series from a postgres database,
that key value pair storage (hstore), to archive time series.
After reading the information from the database a standard
R time series object of class 'ts' is built and returned.
}
\author{
Matthias Bannert, Gabriel Bucur
}

