% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTimeseries.R
\name{readTimeSeries}
\alias{readTimeSeries}
\title{Read Time Series From PostgreSQL database}
\usage{
readTimeSeries(series, con, tbl = "timeseries_main", schema = "timeseries",
  env = NULL)
}
\arguments{
\item{series}{character vector of time series keys}

\item{con}{a PostgreSQL connection object}

\item{tbl}{character string denoting the name of the relation that contains ts_key, ts_data, ts_frequency.}

\item{schema}{SQL schema name. Defaults to timeseries.}

\item{env}{environment, optional argument to dump time series directly into an environment. Most often used with globalenv(), which gives all time series directly back to the global env.}
}
\description{
This function reads a time series from a PostgreSQL relation
that uses Postgres' key value pair storage (hstore).
After reading the information from the database a standard
R time series object of class 'ts' is built and returned. Irregular time series return zoo objects.
}
\author{
Matthias Bannert, Gabriel Bucur
}

