% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readTimeseries.R
\name{readTimeSeries}
\alias{readTimeSeries}
\title{Read Time Series From PostgreSQL database}
\usage{
readTimeSeries(series, con, tbl = "v_timeseries_json",
  schema = "timeseries")
}
\arguments{
\item{series}{character vector of series names.}

\item{con}{a PostgreSQL connection object}

\item{tbl}{character string denoting the name of the view
containing the json records.}

\item{schema}{SQL schema name. Defaults to timeseries.}
}
\description{
This function reads a time series from a postgres database,
that uses key value pair storage (hstore), to archive time series.
After reading the information from the database a standard
R time series object of class 'ts' is built and returned.
}
\author{
Matthias Bannert, Gabriel Bucur
}

