% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readTimeSeries}
\alias{readTimeSeries}
\title{Read Time Series From PostgreSQL database}
\usage{
readTimeSeries(series, con = options()$TIMESERIESDB_CON,
  meta_unlocalized = F, meta_localized = F, tbl = "timeseries_main")
}
\arguments{
\item{series}{character representation of the key of the time series}

\item{con}{a PostgreSQL connection object}

\item{meta_unlocalized}{logical look for unlocalized meta information in the database. Defaults to FALSE}

\item{meta_localized}{look for localized meta information: either "no", "all", or specific country abbreviation. defaults to "no".}

\item{tbl}{character string denoting the name of the main time series table
in the PostgreSQL database.}
}
\description{
This function reads a time series from a postgres database,
that key value pair storage (hstore), to archive time series.
After reading the information from the database a standard
R time series object of class 'ts' is built and returned.
}
\author{
Matthias Bannert, Gabriel Bucur
}

