% Generated by roxygen2 (4.0.2): do not edit by hand
\name{searchKVP}
\alias{lookForKey}
\alias{searchKVP}
\title{Search Key-Value Pairs, look for existing keys in an Hstore}
\usage{
searchKVP(key, value, con = options()$TIMESERIESDB_CON,
  hstore = "meta_data", tbl = "meta_data_unlocalized", where = NULL)

lookForKey(key, con = options()$TIMESERIESDB_CON, hstore = "meta_data",
  tbl = "meta_data_unlocalized", where = NULL)
}
\arguments{
\item{key}{character}

\item{value}{in the hstore}

\item{con}{PostgreSQL connection object}

\item{hstore}{name of the hstore column}

\item{tbl}{name of the table to be queried. defaults to  'meta_data_localized'}

\item{where}{character restrict the SQL query by an additional where clause. Defaults to NULL.
E.g.: ts_key LIKE '%ch%'.}
}
\description{
Search hstore key value in PostgreSQL. Very handsome when crawling the database by meta informaiton.
Currently works for non translated meta information.
}

