% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbExplorerUtils.R
\name{searchKeys}
\alias{searchKeys}
\alias{searchKeys.key}
\alias{searchKeys.md}
\alias{searchKeys.set}
\title{Search Keys From Pre-Defined Or Key Based Queries}
\usage{
searchKeys(x, input = NULL, con, ...)

\method{searchKeys}{key}(x, input = NULL, con, ...)

\method{searchKeys}{md}(x, input = NULL, con, ...)

\method{searchKeys}{set}(x, input = NULL, con, ...)
}
\arguments{
\item{x}{query type as character.}

\item{input}{input that triggers action. Defaults to NULL.}

\item{con}{PostgreSQL Connection object.}

\item{...}{addtional arguments to be passed to the methods.}
}
\description{
Search Keys From Pre-Defined Or Key Based Queries
}

