% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storeMetaInformation.R
\name{storeMetaInformation}
\alias{storeMetaInformation}
\alias{storeMetaInformation.tsmeta.list}
\title{Update Meta Information Records}
\usage{
storeMetaInformation(con, meta, schema = "timeseries",
  tbl = "meta_data_unlocalized", locale = NULL, keys = NULL,
  chunksize = 10000)

\method{storeMetaInformation}{tsmeta.list}(con, meta, schema = "timeseries",
  tbl = "meta_data_unlocalized", locale = NULL, keys = NULL,
  chunksize = 10000)
}
\arguments{
\item{con}{a PostgreSQL connection object}

\item{meta}{a tsmeta.list or tsmets.dt object}

\item{schema}{character name of the schema to write to. Defaults to 'timeseries'.}

\item{tbl}{character name of the meta information table to write to. 
Defaults to 'meta_data_unlocalized'.}

\item{locale}{character iso 2 digit locae description. Defaults to NULL.}

\item{keys}{character vector of time series. If specified only the selected 
meta information is stored. Defaults to NULL which stores all meta information
records in the environment.}

\item{chunksize}{integer max size of chunks to split large query in.}
}
\description{
When a time series is stored to the database by \code{\link{storeTimeSeries}} 
a minimal unlocalized (i.e. untranslatable) meta information record is being 
generated. This meta information can be supplement using the storeMetaInformation
method.
}
