% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transactionUtils.R
\name{beginTransaction}
\alias{beginTransaction}
\title{Convenience Wrapper to SQL classics for BEGIN,COMMIT, ROLLBACK}
\usage{
beginTransaction(con)
}
\arguments{
\item{con}{PostgreSQL connection object.}
}
\description{
this set of function can speed up loops by starting a transaction,
performing several queries and ending them with either commit or rollback.
}

